pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "siseli-bridge"
version = "2.5.12"
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt>=1.6,<2.0",
    "scapy>=2.5,<2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov>=4",
]

[tool.pytest.ini_options]
testpaths = ["siseli_bridge/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["siseli_bridge/src"]
omit = ["*/__pycache__/*", "*/tests/*"]
